Update Code Day 20/05

+++ travel/Models/BookingTour.cs
@@ -8,9 +8,9 @@ namespace travel.Models
     public class BookingTour
     {
         public long Id { get; set; }
-        public long CountAdult { get; set; }
-        public long CountChild { get; set; }
-        public long Price { get; set; } //Giá ở đây là tổng số tiền mà khách phải trả
+        public int CountAdult { get; set; }
+        public int CountChild { get; set; }
+        public double Price { get; set; } //Giá ở đây là tổng số tiền mà khách phải trả
         public DateTime DateBook { get; set; }
         public string UserId { get; set; }
         public virtual ApplicationUser User { get; set; }
